requirements.txt
darkdetect==0.8.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sv-ttk==2.6.0
tzdata==2025.2
